rename and move init function
- Renomear a função para `signInOrSignUp` já que ela é responsável pelo login ou criação da conta
- Mover a função para o próprio arquivo com contexto mais coerente
